move mock data fixtures to utils and sorting test
