--- frontend/src/__tests__/sorting.test.js
@@ -1,39 +1,0 @@
-import sortSaleList from '../utils/sortSaleList';
-import mockData from './data/mockData';
-import { InventoryCard } from '../utils/ScryfallCard';
-
-/**
- * Shuffles an array
- * @param {array} arr - The array to ramdomize
- */
-function randomizeArray(arr) {
-    const output = [];
-
-    // Picks a random index from the source array, splices and them pushes it to output
-    while (arr.length) {
-        const randomCard = arr.splice(
-            Math.floor(Math.random() * arr.length),
-            1
-        );
-        output.push(randomCard[0]);
-    }
-
-    return output.map((c) => new InventoryCard(c));
-}
-
-describe('Sort function', () => {
-    test('Sorts the sammple data properly', () => {
-        const random = randomizeArray(mockData);
-        const sorted = sortSaleList(random);
-
-        expect(sorted[0].name).toBe('Archangel Avacyn // Avacyn, the Purifier');
-        expect(sorted[1].name).toBe('Bazaar Trademage');
-        expect(sorted[2].name).toBe('Necropotence');
-        expect(sorted[3].name).toBe('Wheel of Fortune');
-        expect(sorted[4].name).toBe('Colossal Dreadmaw');
-        expect(sorted[5].name).toBe('World Breaker');
-        expect(sorted[6].name).toBe('Sigarda, Host of Herons');
-        expect(sorted[7].name).toBe('Emrakul, the Aeons Torn');
-        expect(sorted[8].name).toBe("Gaea's Cradle");
-    });
-});
